user-service/requirements.txt
Flask==3.0.3
pymongo==3.11.4
PyJWT==2.7.0
pytest==8.3.3